First Project Of Patika Unity Junior Programmer
We have a car in our game. If we press the up arrow key, it moves forward, and if we press the down arrow key, it moves backward.  When we press the right arrow key, it goes to the right, and when we press the left arrow key, it goes to the left. The game ends when the car hits the last obstacle, which is a barrel. This is the first project of the Patika Unity Junior Programmer course.